Resolve all routes
