+++ src/App.js
@@ -1,13 +1,13 @@
 import { Routes, Route } from 'react-router-dom';
-import Main from './components/Main';
-import EventDetails from './pages/events/EventDetailsPages';
+import EventDetails from './pages/events/EventDetails';
 import Layout from './components/Layout';
+import Index from './pages/events';
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
-        <Route path="/" element={<Main />} />
+        <Route path="/" element={<Index />} />
         <Route path="/events/:id" element={<EventDetails />} />
       </Route>
     </Routes>
