+++ src/pages/events/EventDetails.jsx
